requirements.txt
absl==0.0
absl_py==0.12.0
autodp==0.2
ax==0.52.0
ax_platform==0.1.20
core==1.0.1
easydict==1.9
ipdb==0.13.9
matplotlib==3.1.0
numpy==1.16.4
pandas==0.24.2
Pillow==9.0.1
pretrainedmodels==0.7.4
progress==1.5
pycocotools==2.0.4
PyYAML==6.0
scikit_learn==1.0.2
scipy==1.3.0
seaborn==0.9.0
six==1.12.0
tensorboardX==2.4.1
torch==1.8.0+cu101
torchsummary==1.5.1
torchvision==0.9.0+cu101
tqdm==4.32.1
